mv generateRepoMetadata to java/repometadata.go
